% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_count_species_by_site.R
\name{fb_count_species_by_site}
\alias{fb_count_species_by_site}
\title{Count Number of Species per Site}
\usage{
fb_count_species_by_site(site_species)
}
\arguments{
\item{site_species}{a \code{data.frame} with sites in rows and species in columns.
\strong{NOTE}: the first column should be named \strong{\code{"site"}} and indicate site
names. The other columns should be named according to species names.}
}
\value{
A three-column \code{data.frame} with:
\itemize{
\item \code{site}: the name of the site;
\item \code{n_species}: the number of present species;
\item \code{coverage}: the percentage of present species.
}
}
\description{
This function computes for each site the number and proportion of species
present (distribution value higher than 0 and non-NA) compared to all species
provided. For example, a site could contain only 20\% of all species provided.
}
\examples{
species_coverage_by_site <- fb_count_species_by_site(woodiv_site_species)
head(species_coverage_by_site)
}
\seealso{
Other {count functions}:
\code{\link[=fb_count_sites_by_species]{fb_count_sites_by_species()}},
\code{\link[=fb_count_species_by_trait]{fb_count_species_by_trait()}},
\code{\link[=fb_count_traits_by_species]{fb_count_traits_by_species()}}
}
\concept{{count functions}}
