% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_count_species_by_sites.R
\name{fb_count_species_by_sites}
\alias{fb_count_species_by_sites}
\title{Compute species coverage for each site}
\usage{
fb_count_species_by_sites(site_species)
}
\arguments{
\item{site_species}{a \code{data.frame} with sites in rows and species in columns.
The first column should be named \code{"site"} and contain site names. The other
columns should be named according to species names.}
}
\value{
A three-column \code{data.frame} with:
\itemize{
\item \code{site}: the name of the site;
\item \code{n_species}: the number of present species;
\item \code{coverage}: the percentage of present species.
}
}
\description{
For each site computes the percentage of present species (distribution value
higher than 0 and non-NA).
}
\examples{
library("funbiogeo")

data("site_species")

species_coverage_by_site <- fb_count_species_by_sites(site_species)
}
