% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_cwm.R
\name{fb_cwm}
\alias{fb_cwm}
\title{Compute community-weighted means (CWM) of trait values}
\usage{
fb_cwm(site_species, species_traits)
}
\arguments{
\item{site_species}{a \code{data.frame} with sites in rows and species in columns.
\strong{NOTE}: the first column should be named \strong{\code{"site"}} and indicate site
names. The other columns should be named according to species names.}

\item{species_traits}{a \code{data.frame} with species in rows and
traits as columns. \strong{NOTE}: The first column should be named \strong{\code{"species"}}
and contain species names. The other columns should be named according
to trait names.}
}
\value{
A \code{data.frame} with sites in rows and the following variables:
\itemize{
\item \code{site}, the site label,
\item \code{trait}, the trait label as provided in \code{species_traits},
\item and \code{cwm}, the community-weighted means of quantitative traits values.
}
}
\description{
This function returns the community-weighted mean of provided trait values.
It only works with quantitative traits and will warn you otherwise.
It will remove species that either have \code{NA} values in the \code{site_species}
input or \code{NA} values as their trait in the provided trait object.
}
\details{
The community-weighted mean is a site-based trait mean weighted by the
abundance of the species. It can be written with the following equation:

\deqn{
  \text{CWM}_k = \sum_{i = 1}^S p_{ik} \times t_{ik}
}

with \eqn{\text{CWM}_k} the CWM of site k, \eqn{p_{ik}} the relative
abundance of species \eqn{i} in site \eqn{k}, and \eqn{t_{ik}} the trait of
species \eqn{i} in site \eqn{k}.
}
\examples{
site_cwm <- fb_cwm(head(woodiv_site_species), woodiv_traits)
head(site_cwm)
}
