% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_filter_coverage.R
\name{fb_filter_coverage}
\alias{fb_filter_coverage}
\title{Filter sites with a given trait coverage threshold}
\usage{
fb_filter_coverage(site_species, species_traits, coverage_threshold = 1)
}
\arguments{
\item{site_species}{a \code{matrix} or \code{data.frame} with sites in rows and
species in columns. Rows names (sites) and columns names (species) must be
defined.}

\item{species_traits}{a \code{matrix} or \code{data.frame} with species in rows and
traits in columns. Rows names (species) and columns names (traits) must be
defined.}

\item{coverage_threshold}{a numeric of length 1 between 0 and 1. The
percentage trait coverage threshold}
}
\value{
A subset of \code{site_species} with sites covered by X\% of
abundance/coverage considering all provided traits.
}
\description{
...
}
\examples{
library("funbiogeo")

data("site_species")
data("species_traits")

cover <- fb_filter_coverage(site_species, species_traits)
}
