% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_filter_sites_by_species_coverage.R
\name{fb_filter_sites_by_species_coverage}
\alias{fb_filter_sites_by_species_coverage}
\title{Filter sites with a given species coverage threshold}
\usage{
fb_filter_sites_by_species_coverage(
  site_species,
  threshold_species_proportion = 0
)
}
\arguments{
\item{site_species}{a \code{data.frame} with sites in rows and species in columns.
\strong{NOTE}: the first column should be named \strong{\code{"site"}} and indicate site
names. The other columns should be named according to species names.}

\item{threshold_species_proportion}{a numeric of length 1 between 0 and 1.
The threshold of species coverage under which to exclude the sites.}
}
\value{
A subset of \code{site_species} with sites with at least
\code{threshold_species_proportion}\% species present.
}
\description{
Selects sites (rows) for which the proportion of species present
(distribution value higher than 0 and non-NA) is higher than the user-defined
threshold.
}
\examples{
# Get sites with more than 40\% of the species
new_site_species <- fb_filter_sites_by_species_coverage(
  woodiv_site_species,
  threshold_species_proportion = 0.4
)
# There are now only 148 sites
dim(new_site_species)
new_site_species[1:3, 1:4]
}
