% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_filter_sites_by_trait_coverage.R
\name{fb_filter_sites_by_trait_coverage}
\alias{fb_filter_sites_by_trait_coverage}
\title{Filter sites with a given trait coverage threshold}
\usage{
fb_filter_sites_by_trait_coverage(
  site_species,
  species_traits,
  threshold_traits_proportion = 1
)
}
\arguments{
\item{site_species}{a \code{data.frame} with sites in rows and species in columns.
\strong{NOTE}: the first column should be named \strong{\code{"site"}} and indicate site
names. The other columns should be named according to species names.}

\item{species_traits}{a \code{data.frame} with species in rows and
traits as columns. \strong{NOTE}: The first column should be named \strong{\code{"species"}}
and contain species names. The other columns should be named according
to trait names.}

\item{threshold_traits_proportion}{a numeric of length 1 between 0 and 1.
The percentage trait coverage threshold}
}
\value{
A subset of \code{site_species} with sites covered by X\% of
abundance/coverage considering all provided traits.
}
\description{
Select sites (rows of the \code{site_species} object) with all given traits
available for at least the user-defined proportion of species
\code{threshold_traits_proportion}. If a single trait is given, then the threshold
applies to a single trait, if more than one trait is provided, then the
function considers a threshold across all traits taken together.
}
\examples{
# Filter all the sites where all species have known traits
new_site_species <- fb_filter_sites_by_trait_coverage(
  woodiv_site_species, woodiv_traits
)

# There is only one such site
nrow(new_site_species)

# Filter sites where at least 80\% of species have known traits
new_site_species_2 <- fb_filter_sites_by_trait_coverage(
  woodiv_site_species, woodiv_traits, threshold_traits_proportion = 0.8
)

# There are now four sites
nrow(new_site_species_2)
}
