% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_filter_traits_by_species_coverage.R
\name{fb_filter_traits_by_species_coverage}
\alias{fb_filter_traits_by_species_coverage}
\title{Filter traits with a given species coverage threshold}
\usage{
fb_filter_traits_by_species_coverage(
  species_traits,
  threshold_species_proportion = 0
)
}
\arguments{
\item{species_traits}{a \code{data.frame} with species in rows and
traits as columns. \strong{NOTE}: The first column should be named \strong{\code{"species"}}
and contain species names. The other columns should be named according
to trait names.}

\item{threshold_species_proportion}{\code{numeric(1)} [default = \code{NULL}]\cr{}
between 0 and 1. The percentage of species coverage threshold.}
}
\value{
A subset of \code{species_traits} with traits for the specified
proportion of species.
}
\description{
Selects traits (columns) for which the percentage of species without
\code{NA} (missing trait values) is higher than a user-defined threshold.
}
\examples{
# Filter traits that have at least 60\% non-missing values
new_species_traits <- fb_filter_traits_by_species_coverage(
  woodiv_traits,
  threshold_species_proportion = 0.6
)

# There are now only 2 traits
ncol(new_species_traits)
}
