% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_format_sites_locations.R
\name{fb_format_sites_locations}
\alias{fb_format_sites_locations}
\title{Format sites x locations object}
\usage{
fb_format_sites_locations(data, site, longitude, latitude, na_rm = FALSE)
}
\arguments{
\item{data}{a \code{data.frame} in a long format (see example).}

\item{site}{a character of length 1. Name of the column with site labels.}

\item{longitude}{a character of length 1. Name of the column with longitude.}

\item{latitude}{a character of length 1. Name of the column with latitude.}

\item{na_rm}{a logical value. If \code{TRUE} remove sites with incomplete
coordinates. Default is \code{FALSE}.}
}
\value{
A matrix with sites in rows and longitude and latitude in columns.
}
\description{
...
}
\examples{
library("funbiogeo")

filename <- system.file("extdata", "raw_trees_data.csv", 
                        package = "funbiogeo")
all_data <- read.csv2(filename)

head(all_data)

sites_locations <- fb_format_sites_locations(all_data, "site", "longitude", 
                                             "latitude")
head(sites_locations)
}
