% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_format_sites_species.R
\name{fb_format_sites_species}
\alias{fb_format_sites_species}
\title{Format sites x species object}
\usage{
fb_format_sites_species(data, site, species, value, na_to_zero = TRUE)
}
\arguments{
\item{data}{a \code{data.frame} in a long format (see example).}

\item{site}{a character of length 1. Name of the column with site labels.}

\item{species}{a character of length 1. Name of the column with species
names.}

\item{value}{a character of length 1. Name of the column with species
occurrence/abundance.}

\item{na_to_zero}{a logical value. If \code{TRUE} (default) NA are replaced by
\code{0}.}
}
\value{
A matrix with sites in rows and species in columns.
}
\description{
...
}
\examples{
library("funbiogeo")

filename <- system.file("extdata", "raw_trees_data.csv", 
                        package = "funbiogeo")
all_data <- read.csv2(filename)

head(all_data)

sites_species <- fb_format_sites_species(all_data, "site", "species", 
                                         "count")
head(sites_species)
}
