% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_format_species_traits.R
\name{fb_format_species_traits}
\alias{fb_format_species_traits}
\title{Extract species x traits information from long format data}
\usage{
fb_format_species_traits(data, species, traits)
}
\arguments{
\item{data}{a \code{data.frame} in a long format (see example).}

\item{species}{a \code{character} of length 1. Name of the column with species
names.}

\item{traits}{a \code{character} of length >= 1. Name(s) of trait column(s).}
}
\value{
A \code{data.frame} with species in rows and traits in columns, with the
first column names \code{"species"} containing the species names.
}
\description{
Convert a flat \code{data.frame} with traits values for different species
into a proper \code{species_traits} object that can then be used by other
functions. The final output contains species in rows and traits in columns.
}
\examples{
filename <- system.file(
  "extdata", "woodiv_raw_data.csv", package = "funbiogeo"
)
all_data <- read.csv(filename)

head(all_data)

traits <- c("plant_height", "seed_mass", "sla", "wood_density")

species_traits <- fb_format_species_traits(all_data, "species", traits)
head(species_traits)
}
