% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_get_all_coverages.R
\name{fb_get_all_coverages}
\alias{fb_get_all_coverages}
\title{Compute Trait Coverage per Site for All Traits}
\usage{
fb_get_all_coverages(site_species, species_traits)
}
\arguments{
\item{site_species}{a \code{data.frame} with sites in rows and species in columns.
The first column should be named \code{"site"} and contain site names. The other
columns should be named according to species names.}

\item{species_traits}{a \code{data.frame} with species in rows and
traits as columns. The first column should be named \code{"species"}
and contain species names. The other columns should be named according
to trait names.}
}
\value{
a data.frame with a column with sites and one column per provided
trait giving its coverage (percent species per site, weighted by abundance
that have trait data)
}
\description{
Compute Trait Coverage per Site for All Traits
}
\examples{
fb_get_all_coverages(site_species, species_traits)
}
