% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_get_coverage.R
\name{fb_get_coverage}
\alias{fb_get_coverage}
\title{Compute traits coverage}
\usage{
fb_get_coverage(sites_species, species_traits)
}
\arguments{
\item{sites_species}{a \code{matrix} or \code{data.frame} with sites in rows and
species in columns. Rows names (sites) and columns names (species) must be
defined.}

\item{species_traits}{a \code{matrix} or \code{data.frame} with species in rows and
traits in columns. Rows names (species) and columns names (traits) must be
defined.}
}
\value{
A \code{data.frame} with \code{n} rows (where \code{n} is the number of sites) and
two columns: \code{site}, the site label, and \code{trait_coverage}, the percent of
total abundance/presence of species that have traits data.
}
\description{
Compute traits coverage for all sites, i.e., the percentage of total
abundance/presence of species that have traits data compared to total
species.
This function assumes that all species provided in the traits dataset have
all their traits specified.
}
\examples{
library("funbiogeo")

data("species_occs")
data("species_traits")

cover <- fb_get_coverage(species_occs, species_traits)
}
