% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_get_environment.R
\name{fb_get_environment}
\alias{fb_get_environment}
\title{Extract Raster Values at Location of Sites}
\usage{
fb_get_environment(site_locations, environment_raster)
}
\arguments{
\item{site_locations}{a \code{sf} object with the spatial geometries of sites.
\strong{NOTE}: the first column should be named \strong{\code{"site"}}
and indicate site names.}

\item{environment_raster}{a \code{SpatRaster} object (package \code{terra}).
A single or multi-layers environmental raster.}
}
\value{
A \code{data.frame} with average environmental values (columns) per site
(rows), with the first column being \code{"site"}, indicating site names.
}
\description{
This function uses a \code{site_locations} object and a given \code{SpatRaster}.
It extracts the mean value of the provided raster for each site.
}
\examples{
## Import climate rasters ----
prec <- system.file("extdata", "annual_tot_prec.tif", package = "funbiogeo")
tavg <- system.file("extdata", "annual_mean_temp.tif", package = "funbiogeo")

layers <- terra::rast(c(tavg, prec))

fb_get_environment(head(woodiv_locations), layers)
}
