% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_get_trait_coverage_by_site.R
\name{fb_get_trait_coverage_by_site}
\alias{fb_get_trait_coverage_by_site}
\title{Compute Trait Coverage For Each Site Weighted by Abundance}
\usage{
fb_get_trait_coverage_by_site(site_species, species_traits)
}
\arguments{
\item{site_species}{a \code{data.frame} with sites in rows and species in columns.
\strong{NOTE}: the first column should be named \strong{\code{"site"}} and indicate site
names. The other columns should be named according to species names.}

\item{species_traits}{a \code{data.frame} with species in rows and
traits as columns. \strong{NOTE}: The first column should be named \strong{\code{"species"}}
and contain species names. The other columns should be named according
to trait names.}
}
\value{
A \code{data.frame} with \code{n} rows (where \code{n} is the number of sites) and
two columns: \code{site}, the site label, and \code{trait_coverage}, the percent of
total abundance/presence of species that have traits data.
}
\description{
Compute trait coverage for all sites, i.e., the proportion of total
abundance/presence of species that have traits data compared to all species.
This function assumes that all species provided in the trait dataset have
all their traits specified (meaning that all species have either known or
\code{NA} values reported as their traits).
\strong{NB}: this function returns trait coverage using all traits
provided in the input \code{species_traits} data.frame.
}
\examples{
site_trait_cov <- fb_get_trait_coverage_by_site(
    woodiv_site_species, woodiv_traits
)

head(site_trait_cov)
}
