% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_map_raster.R
\name{fb_map_raster}
\alias{fb_map_raster}
\title{Map a Single Raster Layer}
\usage{
fb_map_raster(x, ...)
}
\arguments{
\item{x}{a \code{SpatRaster} object (package \code{terra}). A raster of one single
layer}

\item{...}{other parameters passed to \code{ggplot2::theme()}}
}
\value{
A \code{ggplot} object.
}
\description{
Map a Single Raster Layer
}
\examples{
library(ggplot2)

## Load raster ----
tavg <- system.file("extdata", "annual_mean_temp.tif", package = "funbiogeo")
tavg <- terra::rast(tavg)

## Default map ----
fb_map_raster(tavg)

## Map with custom theme ----
fb_map_raster(tavg, legend.position = "bottom")

## Advanced customization ----
my_map <- fb_map_raster(tavg) + 
  scale_fill_distiller("Temperature", palette = "Spectral") +
  theme(legend.position = "bottom") + 
  ggtitle("Mean annual temperature in Western Europe")

my_map

## Map w/o annotation ----
fb_map_raster(tavg) + 
  theme_void() + 
  theme(legend.position = "none")
}
