% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_plot_number_species_by_trait.R
\name{fb_plot_number_species_by_trait}
\alias{fb_plot_number_species_by_trait}
\title{Show Number of Species per Trait}
\usage{
fb_plot_number_species_by_trait(
  species_traits,
  threshold_species_number = NULL
)
}
\arguments{
\item{species_traits}{a \code{data.frame} with species in rows and
traits as columns. The first column should be named \code{"species"}
and contain species names. The other columns should be named according
to trait names.}

\item{threshold_species_number}{\code{NULL} or \code{numeric(1)} [default = \code{NULL}]\cr{}
If \code{NULL} doesn't show a line otherwise adds a vertical line
at the specified number of species}
}
\value{
a ggplot2 object
}
\description{
Show Number of Species per Trait
}
\examples{
data(species_traits)

\dontrun{\%
fb_plot_number_species_by_trait(species_traits)

# Add a vertical cutoff line
fb_plot_number_species_by_trait(species_traits, 100)
}

}
