% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_plot_number_traits_by_species.R
\name{fb_plot_number_traits_by_species}
\alias{fb_plot_number_traits_by_species}
\title{Plot Number of Traits per Species}
\usage{
fb_plot_number_traits_by_species(
  species_traits,
  species_categories = NULL,
  threshold_species_proportion = NULL
)
}
\arguments{
\item{species_traits}{a \code{data.frame} with species in rows and
traits as columns. \strong{NOTE}: The first column should be named \strong{\code{"species"}}
and contain species names. The other columns should be named according
to trait names.}

\item{species_categories}{(default = \code{NULL}) 2-columns \code{data.frame} giving
species categories, with the first column describing the species name, and
the second column giving their corresponding categories}

\item{threshold_species_proportion}{\code{numeric(1)} [default = \code{NULL}]\cr{}
between 0 and 1. The percentage of species coverage threshold.}
}
\value{
a \code{ggplot2} object
}
\description{
Display a graph showing the number (and proportion) of species having
at least 0, 1, etc. number of traits. It provides a lollipop graph to examine
which number of trait cover a certain proportion of the species.
This plot doesn't show which traits are concerned.
}
\examples{
fb_plot_number_traits_by_species(woodiv_traits)

# Add a vertical cutoff line (33\% of the species)
fb_plot_number_traits_by_species(
    woodiv_traits, threshold_species_proportion = 1/3
)

}
