% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_plot_site_environment.R
\name{fb_plot_site_environment}
\alias{fb_plot_site_environment}
\title{Plot Position of Sites in Environmental Space}
\usage{
fb_plot_site_environment(
  site_locations,
  environment_raster,
  first_layer = names(environment_raster)[1],
  second_layer = names(environment_raster)[2]
)
}
\arguments{
\item{site_locations}{a \code{sf} object with the spatial geometries of sites.
\strong{NOTE}: the first column should be named \strong{\code{"site"}}
and indicate site names.}

\item{environment_raster}{a \code{SpatRaster} object (package \code{terra}).
A single or multi-layers environmental raster.}

\item{first_layer}{\code{character(1)} the name of the first layer to use,
by default uses the first layer of \code{environment_raster}}

\item{second_layer}{\code{character(1)} the name of the second layer to use,
by default uses the second layer of \code{environment_raster}}
}
\value{
a \code{ggplot} object
}
\description{
Plot a figure showing the average environmental space of given sites compared
to a full environmental vector. For the sake of simplicity only represents
the figure along two environmental axes. The average environmental value are
extracted for each site.
}
\examples{
# Import climate rasters
prec <- system.file("extdata", "annual_tot_prec.tif", package = "funbiogeo")
tavg <- system.file("extdata", "annual_mean_temp.tif", package = "funbiogeo")

layers <- terra::rast(c(tavg, prec))

# Make plot (show environmental position of 6 first sites)
fb_plot_site_environment(head(woodiv_locations), layers)
}
