% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_plot_trait_correlation.R
\name{fb_plot_trait_correlation}
\alias{fb_plot_trait_correlation}
\title{Plot Trait Correlation Matrix}
\usage{
fb_plot_trait_correlation(species_traits, ...)
}
\arguments{
\item{species_traits}{a \code{data.frame} with species in rows and
traits as columns. \strong{NOTE}: The first column should be named \strong{\code{"species"}}
and contain species names. The other columns should be named according
to trait names.}

\item{...}{Additional options passed to \code{\link[stats:cor]{stats::cor()}}}
}
\value{
a \code{ggplot} object
}
\description{
This functions outputs a \code{ggplot2} figure that outputs a trait correlation
matrix. It only works on numerical traits and will output messages if
non-numerical traits are found and errors if no numerical traits are found.
Internally it uses the \code{\link[stats:cor]{stats::cor()}} function and only works on complete
observation (it removes any row that contains any \code{NA}).
Use the \code{...} argument to pass options to the \code{cor()} function.
}
\examples{
fb_plot_trait_correlation(species_traits)

# Plot Spearman's correlation
fb_plot_trait_correlation(species_traits, method = "spearman")
}
