% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_table_trait_summary.R
\name{fb_table_trait_summary}
\alias{fb_table_trait_summary}
\title{Summary Table on Traits (Missingness, Range, etc.)}
\usage{
fb_table_trait_summary(species_traits, kable = FALSE)
}
\arguments{
\item{species_traits}{a \code{data.frame} with species in rows and
traits as columns. \strong{NOTE}: The first column should be named \strong{\code{"species"}}
and contain species names. The other columns should be named according
to trait names.}

\item{kable}{\code{TRUE} or \code{FALSE} Should function returns a \code{\link[knitr:kable]{knitr::kable()}}?
defaults \code{FALSE}}
}
\value{
a \code{data.frame} with the following columns:
\itemize{
\item \code{trait_name}: a \code{character} column with the trait name as indicated in
\code{species_traits}
\item \code{trait_type}: the nature of the trait (\code{numeric}, \code{categorical},
or \code{ordered})
\item \code{number_non_missing}: the total number of non-\code{NA} trait values
\item \code{proportion_non_missing}: the proportion of non-\code{NA} trait values
\item \code{trait_range}: for numerical traits, the range of values
\item \code{trait_mean_sd}: for numerical traits, the mean plus-minus the standard
deviation
\item \code{number_distinct}: for non-numerical traits, the number of categories
\item \code{list_distinct}: for non-numerical traits, the list of categories
}
}
\description{
This function outputs a \code{data.frame} that summarises the species by trait
table to have many information in a glance. This can then return a data.frame
or a nicely formatted \code{\link[knitr:kable]{knitr::kable()}} for inclusion in an Rmarkdown
document.
}
\examples{
fb_table_trait_summary(species_traits)
}
