% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_upscale_sites.R
\name{fb_upscale_sites}
\alias{fb_upscale_sites}
\title{Upscale sites}
\usage{
fb_upscale_sites(
  sites_locations,
  data,
  grid,
  fun = mean,
  crs = "+proj=longlat +datum=WGS84 +no_defs"
)
}
\arguments{
\item{sites_locations}{a \code{matrix} or \code{data.frame} with sites in rows and two
columns: longitude and latitude. The first column must be the longitude.}

\item{data}{a \code{matrix} or \code{data.frame} containing values to rasterize. Can
have one or several columns (variables to rasterize). Row names must
contain sites names as in the argument \code{sites_locations}.}

\item{grid}{a \code{SpatRaster} object (package \code{terra}). A raster of one single
layer.}

\item{fun}{the function used to aggregate points values when there are
multiple points in one cell. Default is \code{mean}.}

\item{crs}{a character of length 1 specifying the Coordinate Reference
System in the PROJ4 standard.
Default is: \code{'+proj=longlat +datum=WGS84 +no_defs'}.}
}
\value{
A \code{SpatRaster} object with \code{n} layers where \code{n} represents the number
of columns in \code{data}.
}
\description{
Upscale sites
}
\examples{
library("funbiogeo")

data("sites_locs")
data("species_occs")

## Import grid ----
tavg <- system.file("extdata", "annual_mean_temp.tif", package = "funbiogeo")
tavg <- terra::rast(tavg)

## Rasterize 3 first species counts ----
rasters <- fb_upscale_sites(sites_locations = sites_locs, 
                            data            = species_occs[ , 1:3], 
                            grid            = tavg)

## Subset first layer ----
a_neg <- terra::subset(rasters, "acer_negundo")

## Species maps ----
fb_map_raster(a_neg) + 
  ggplot2::scale_fill_distiller("Counts", palette = "Blues", direction = 1) +
  ggplot2::ggtitle("Acer negundo in Pennsylvania")
}
