% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_species_traits_completeness.R
\name{plot_species_traits_completeness}
\alias{plot_species_traits_completeness}
\title{Show Species by Trait Completeness}
\usage{
plot_species_traits_completeness(species_traits)
}
\arguments{
\item{species_traits}{a \code{data.frame} with species in rows and
traits as columns. The first column should be named \code{"species"}
and contain species names. The other columns should be named according
to trait names.}
}
\value{
a \code{ggplot2} object
}
\description{
Show Species by Trait Completeness
}
\examples{
data(species_traits)
plot_species_traits_completeness(species_traits)
}
