% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{site_locations}
\alias{site_locations}
\title{Data: sites x locations of Western Europe mammals occurrences}
\format{
An \code{sf} object with 1,505 rows (sites) with a WGS84 (EPSG:4236)
Coordinate Reference System
}
\usage{
site_locations
}
\description{
This dataset contains spatial coordinates of sites of Western Europe where
species presence/absence (see \code{site_species} dataset) has been sampled.

See https://github.com/frbcesab/eumammals for a complete description of the
workflow used to create this dataset.

This dataset shows the format of the argument \code{site_locations} used in
several functions of \code{funbiogeo}.

Note that sites labels are specified in row names.
}
\examples{
data("site_locations")
class(site_locations)
dim(site_locations)
head(site_locations)
}
\keyword{datasets}
