% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sites_locs}
\alias{sites_locs}
\title{Coordinates of Pennsylvania survey sites}
\format{
A data frame with 5,770 rows (sites) and the following two columns:
\describe{
\item{longitude}{longitude of the site (in WGS84)}
\item{latitude}{latitude of the site (in WGS84)}
}
}
\usage{
sites_locs
}
\description{
This dataset contains spatial coordinates of sites sampled in Pennsylvania
where trees species were counted (individuals counts). This dataset shows
the format of the argument \code{sites_locations} used in several functions of
\code{funbiogeo}. Note that sites labels are specified in row names.
}
\examples{
data("sites_locs")
head(sites_locs)
}
\keyword{datasets}
