% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{species_categories}
\alias{species_categories}
\title{Species x Categories of Western Europe Mammals}
\format{
A \code{data.frame} with 149 rows (species) and 2 columns (1 column for
species name and 1 for the order).
}
\usage{
species_categories
}
\description{
This dataset contains the order of mammals species in Western
Europe. This dataset is derived from IUCN range maps downloaded at:
\url{https://www.iucnredlist.org/resources/spatial-data-download}. In
respect to the license, species names have been anonymized.
}
\details{
See \url{https://github.com/frbcesab/eumammals} for a complete description
of the workflow used to create this dataset.

This dataset shows the format of the argument \code{species_categories} used in
several functions of \code{funbiogeo}.

Note that species names (\code{species}) is the first column of this dataset.
}
\examples{
data("species_categories")
class(species_categories)
dim(species_categories)
species_categories[1:6, ]
}
\references{
IUCN (2021) The IUCN Red List of Threatened Species. Version 2021-2.
Downloaded on 2021/10/11.
}
\keyword{datasets}
