% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{species_occs}
\alias{species_occs}
\title{Tree species counts in Pennsylvania survey sites}
\format{
A matrix with 5,770 rows (sites) and 93 columns (species).
}
\usage{
species_occs
}
\description{
This dataset contains trees species counts in sites sampled in Pennsylvania.
This dataset shows the format of the argument \code{site_species} used in
several functions of \code{funbiogeo}. Note that sites labels are specified in
row names.
}
\examples{
data("species_occs")
class(species_occs)
dim(species_occs)
species_occs[1:6, 1:6]
}
\keyword{datasets}
