% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{species_traits}
\alias{species_traits}
\title{Tree species traits}
\format{
A matrix with 67 rows (species) and the following three columns:
\describe{
\item{sla}{Specific Leaf Area}
\item{n_mass}{Mass}
\item{wood_dens}{Wood density}
}
}
\usage{
species_traits
}
\description{
This dataset contains traits of trees species. It shows the format of the
argument \code{species_traits} used in several functions of \code{funbiogeo}. Note
that species names are specified in row names.
}
\examples{
data("species_traits")
class(species_traits)
dim(species_traits)
head(species_traits)
}
\keyword{datasets}
