% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{woodiv_categories}
\alias{woodiv_categories}
\title{Species x Categories of some Mediterranean Conifers}
\format{
A \code{data.frame} with 24 rows (species) and 6 columns (\code{species},
\code{family}, \code{genus}, \code{binomial}, \code{endemism}, and \code{cultivated}).
}
\usage{
woodiv_categories
}
\description{
This dataset is derived from the WOODIV database (available at:
\url{https://www.nature.com/articles/s41597-021-00873-3}). It contains
information (classification, endemism, etc.) of the 24 Conifer tree species
occurring in Portugal, Spain, France, and Italy (Mediterranean part).
This information will be used as a species grouping factor.

This dataset shows the format of the argument \code{species_categories} used in
several functions of \code{funbiogeo}. Note that species names must be
the first column and must be named \code{species}.
}
\examples{
data("woodiv_categories")
class(woodiv_categories)
dim(woodiv_categories)
woodiv_categories[1:6, ]
}
\references{
Monnet AC, Cilleros K, Médail F \emph{et al.} (2021) WOODIV, a database of
occurrences, functional traits, and phylogenetic data for all
Euro-Mediterranean trees. \strong{Scientific Data}, 8, 89.
DOI: \doi{10.1038/s41597-021-00873-3}.
}
\keyword{datasets}
