% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_aggregate_site_data.R
\name{fb_aggregate_site_data}
\alias{fb_aggregate_site_data}
\title{Aggregate Site Data Along Coarser Spatial Grid}
\usage{
fb_aggregate_site_data(site_locations, site_data, agg_geom, fun = mean, ...)
}
\arguments{
\item{site_locations}{an \code{sf} object with the spatial geometries of sites.
\strong{NOTE}: the first column should be named \strong{\code{"site"}}
and indicate site names.}

\item{site_data}{a \code{matrix} or \code{data.frame} containing values per sites to
aggregate along the provided grid. Can have one or several columns
(variables to aggregate). The first column must contain sites names as
provided in the first argument \code{site_locations}.}

\item{agg_geom}{a \code{terra::SpatRaster} or an \code{sf} object. This defines the
geometry along which to aggregate the initial data. See more in the Details
section.}

\item{fun}{the function used to aggregate points values when there are
multiple points in one cell. Default is \code{mean}.}

\item{...}{additional argument(s) passed to the provided function \code{fun}}
}
\value{
An object of the same type as the \code{agg_geom} input with as many
layers (if \code{SpatRaster}) or columns (if \code{sf}) as columns provided in the
input \code{site_data}.
}
\description{
This function helps aggregating site data along a coarser grid.
The idea is that you have any type of data at the site scale
(diversity metrics, environmental data, etc.) but you would like to work
on it at a coarser scale, or you want to visualize it at that scale.
This function helps you do exactly that.
}
\details{
The \code{agg_geom} object will condition the type of object ouput by the
function. It can be of any sort as an \code{SpatRaster} or \code{sf} object. Depending
on the need, it could be a regular square grid or hexagonal grid, it could
also be irregular polygons like biomes or ecoregions, or points, and even
lines (such as when aggregating across transects or trajectories).
}
\examples{
## Raster grid
tavg <- system.file(
  "extdata", "annual_mean_temp.tif", package = "funbiogeo"
)
tavg <- terra::rast(tavg)

# Rasterize 3 first species counts
fb_aggregate_site_data(
    head(woodiv_locations), woodiv_site_species[, 1:4], tavg, fun = sum
)

## Irregular polygons
countries <- readRDS(system.file(
  "extdata", "countries_sf.rds", package = "funbiogeo"
))
# Aggregate occurrence per country
fb_aggregate_site_data(
    head(woodiv_locations, n = 20), woodiv_site_species[, 1:4], countries,
    fun = sum
)
}
