% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_aggregate_site_data.R
\name{fb_aggregate_site_data}
\alias{fb_aggregate_site_data}
\title{Aggregate site data along coarser grid}
\usage{
fb_aggregate_site_data(site_locations, site_data, agg_grid, fun = mean)
}
\arguments{
\item{site_locations}{an \code{sf} object with all sites.}

\item{site_data}{a \code{matrix} or \code{data.frame} containing values per sites to
aggregate along the provided grid. Can have one or several columns
(variables to aggregate). The first column must contain sites names as
provided in the example dataset \code{site_locations}.}

\item{agg_grid}{a \code{SpatRaster} object (package \code{terra}).
A raster of one single layer, that defines the grid along which
to aggregate.}

\item{fun}{the function used to aggregate points values when there are
multiple points in one cell. Default is \code{mean}.}
}
\value{
A \code{SpatRaster} object with as many layers as columns in \code{site_data}.
}
\description{
This function helps aggregating site data along a coarser grid.
The idea is that you have any type of data at the site scale
(diversity metrics, environmental data, etc.) but you would like to work
on it at a coarser scale, or you want to visualize it at that scale.
This function helps you do exactly that.
}
\examples{
library("funbiogeo")

data("site_locations")
data("site_species")

## Import grid ----
tavg <- system.file("extdata", "annual_mean_temp.tif", package = "funbiogeo")
tavg <- terra::rast(tavg)

## Rasterize 3 first species counts ----
}
