% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_aggregate_site_data.R
\name{fb_aggregate_site_data}
\alias{fb_aggregate_site_data}
\title{Upscale sites: aggregates them at higher scales}
\usage{
fb_aggregate_site_data(sites_locations, site_data, agg_grid, fun = mean)
}
\arguments{
\item{sites_locations}{an \code{sf} object with all sites.}

\item{site_data}{a \code{matrix} or \code{data.frame} containing values per sites to
aggregate along the provided grid. Can have one or several columns
(variables to aggregate). Row names must  contain sites names as provided
as \code{sites_locations}.}

\item{agg_grid}{a \code{SpatRaster} object (package \code{terra}).
A raster of one single layer, that defines the grid along which
to aggregate.}

\item{fun}{the function used to aggregate points values when there are
multiple points in one cell. Default is \code{mean}.}
}
\value{
A \code{SpatRaster} object with as many layers as columns in \code{site_data}.
}
\description{
This function helps upscaling site data. In our sense upscaling is when
you want to work with a version of your data aggregated at a coarser spatial
scale. This function aggregates the site data at a given scale.
}
\examples{
library("funbiogeo")

data("sites_locs")
data("species_occs")

## Import grid ----
tavg <- system.file("extdata", "annual_mean_temp.tif", package = "funbiogeo")
tavg <- terra::rast(tavg)

## Rasterize 3 first species counts ----
}
