% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_count_sites_by_species.R
\name{fb_count_sites_by_species}
\alias{fb_count_sites_by_species}
\title{Count Number of Sites Occupied by Species}
\usage{
fb_count_sites_by_species(site_species)
}
\arguments{
\item{site_species}{a \code{data.frame} with sites in rows and species in columns.
\strong{NOTE}: the first column should be named \strong{\code{"site"}} and indicate site
names. The other columns should be named according to species names.}
}
\value{
A three-column \code{data.frame} with:
\itemize{
\item \code{species}: the name of the species;
\item \code{n_sites}: the number of sites where the species is present;
\item \code{coverage}: the percentage of sites where the species is present.
}
}
\description{
This function computes the number and proportion of sites occupied by each
species (distribution value higher than 0 and non-NA).
}
\examples{
site_coverage_by_species <- fb_count_sites_by_species(woodiv_site_species)
head(site_coverage_by_species)
}
