% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_count_traits_by_species.R
\name{fb_count_traits_by_species}
\alias{fb_count_traits_by_species}
\title{Compute Number of Known Trait(s) per Species}
\usage{
fb_count_traits_by_species(species_traits)
}
\arguments{
\item{species_traits}{a \code{data.frame} with species in rows and
traits as columns. \strong{NOTE}: The first column should be named \strong{\code{"species"}}
and contain species names. The other columns should be named according
to trait names.}
}
\value{
A three-column \code{data.frame} with:
\itemize{
\item \code{species}: the name of the species;
\item \code{n_traits}: the number of traits with non-missing value for the species;
\item \code{coverage}: the percentage of traits with non-missing value for the
species.
}
}
\description{
For each species, this function computes the number and proportion of traits
without \code{NA} (missing trait values).
}
\examples{
trait_coverage_by_species <- fb_count_traits_by_species(woodiv_traits)
head(trait_coverage_by_species)
}
\seealso{
Other {count functions}:
\code{\link[=fb_count_sites_by_species]{fb_count_sites_by_species()}},
\code{\link[=fb_count_species_by_site]{fb_count_species_by_site()}},
\code{\link[=fb_count_species_by_trait]{fb_count_species_by_trait()}}
}
\concept{{count functions}}
