% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_cwm.R
\name{fb_cwm}
\alias{fb_cwm}
\title{Compute community-weighted means of trait values}
\usage{
fb_cwm(site_species, species_traits)
}
\arguments{
\item{site_species}{a \code{matrix} or \code{data.frame} with sites in rows and
species in columns. Rows names (sites) and columns names (species) must be
defined.}

\item{species_traits}{a \code{matrix} or \code{data.frame} with species in rows and
traits in columns. Rows names (species) and columns names (traits) must be
defined.}
}
\value{
A \code{data.frame} with sites in rows and the following variables:
\code{site}, the site label, \code{trait}, quantitative traits, and \code{cwm}, the
community-weighted means of quantitative traits values.
}
\description{
...
}
\examples{
library("funbiogeo")

data("site_species")
data("species_traits")

cover <- fb_cwm(site_species, species_traits)
}
