% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_filter_sites_by_trait_coverage.R
\name{fb_filter_sites_by_trait_coverage}
\alias{fb_filter_sites_by_trait_coverage}
\title{Filter sites with a given trait coverage threshold}
\usage{
fb_filter_sites_by_trait_coverage(
  site_species,
  species_traits,
  threshold_traits_proportion = 1
)
}
\arguments{
\item{site_species}{a \code{data.frame} with sites in rows and species in columns.
\strong{NOTE}: the first column should be named \strong{\code{"site"}} and indicate site
names. The other columns should be named according to species names.}

\item{species_traits}{a \code{data.frame} with species in rows and
traits as columns. \strong{NOTE}: The first column should be named \strong{\code{"species"}}
and contain species names. The other columns should be named according
to trait names.}

\item{threshold_traits_proportion}{a numeric of length 1 between 0 and 1.
The percentage trait coverage threshold}
}
\value{
A subset of \code{site_species} with sites covered by X\% of
abundance/coverage considering all provided traits.
}
\description{
...
}
\examples{
library("funbiogeo")

data("site_species")
data("species_traits")

cover <- fb_filter_sites_by_trait_coverage(site_species, species_traits)
}
