% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_filter_species_by_site_coverage.R
\name{fb_filter_species_by_site_coverage}
\alias{fb_filter_species_by_site_coverage}
\title{Filter species with a given sites coverage threshold}
\usage{
fb_filter_species_by_site_coverage(
  site_species,
  threshold_sites_proportion = 0
)
}
\arguments{
\item{site_species}{a \code{data.frame} with sites in rows and species in columns.
\strong{NOTE}: the first column should be named \strong{\code{"site"}} and indicate site
names. The other columns should be named according to species names.}

\item{threshold_sites_proportion}{a numeric of length 1 between 0 and 1.
The percentage of sites coverage threshold.}
}
\value{
A subset of \code{site_species} with species with a prevalence higher
than \code{threshold_sites_proportion}.
}
\description{
Selects species (columns) for which the percentage of sites where the
species is present (distribution value higher than 0 and non-NA) is higher
than a threshold.
}
\examples{
library("funbiogeo")

data("site_species")

site_species <- fb_filter_species_by_site_coverage(
  site_species,
  threshold_sites_proportion = 0.1)
}
