% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_filter_species_by_trait_coverage.R
\name{fb_filter_species_by_trait_coverage}
\alias{fb_filter_species_by_trait_coverage}
\title{Filter species with a given traits coverage threshold}
\usage{
fb_filter_species_by_trait_coverage(
  species_traits,
  threshold_traits_proportion = 0
)
}
\arguments{
\item{species_traits}{a \code{data.frame} with species in rows and
traits as columns. The first column should be named \code{"species"}
and contain species names. The other columns should be named according
to trait names.}

\item{threshold_traits_proportion}{a numeric of length 1 between 0 and 1.
The percentage of traits coverage threshold.}
}
\value{
A subset of \code{species_traits} with species covered by X\% of traits.
}
\description{
Selects species (rows) for which the percentage of traits without
\code{NA} (missing trait values) is higher than a threshold.
}
\examples{
library("funbiogeo")

data("species_traits")

species_traits <- fb_filter_species_by_trait_coverage(
  species_traits,
  threshold_traits_proportion = 0.6)
}
