% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_filter_species_traits.R
\name{fb_filter_traits_by_species_coverage}
\alias{fb_filter_traits_by_species_coverage}
\title{Filter traits with a given species coverage threshold}
\usage{
fb_filter_traits_by_species_coverage(species_traits, threshold = 0)
}
\arguments{
\item{species_traits}{a \code{data.frame} with species in rows and
traits as columns. The first column should be named \code{"species"}
and contain species names. The other columns should be named according
to trait names.}

\item{threshold}{a numeric of length 1 between 0 and 1. The percentage of
species coverage threshold.}
}
\value{
A subset of \code{species_traits} with traits covered by X\% of species.
}
\description{
Selects traits (columns) for which the percentage of species without
\code{NA} (missing trait values) is higher than a threshold.
}
\examples{
library("funbiogeo")

data("species_traits")

species_traits <- fb_filter_traits_by_species_coverage(species_traits,
                                                       threshold = 0.6)
}
