% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_format_site_locations.R
\name{fb_format_site_locations}
\alias{fb_format_site_locations}
\title{Extract site x locations information from long format data}
\usage{
fb_format_site_locations(
  data,
  site,
  longitude,
  latitude,
  crs = sf::st_crs(4326),
  na_rm = FALSE
)
}
\arguments{
\item{data}{a \code{data.frame} in a long format (see example).}

\item{site}{a \code{character} of length 1. Name of the column with site labels.}

\item{longitude}{a \code{character} of length 1. Name of the column with
longitude. The function assumes coordinates are WGS84 (EPSG:4326).}

\item{latitude}{a \code{character} of length 1. Name of the column with latitude.
The function assumes coordinates are WGS84 (EPSG:4326).}

\item{crs}{a \code{character} of length 1 or an object of class \code{crs}.
Coordinate Reference System (CRS) of the specified coordinates.
The CRS should be a \href{https://r.geocompx.org/spatial-class.html?q=CRS#crs-intro}{valid CRS in R}.
It can either be a character like \code{"+proj=longlat +datum=WGS84 +no_defs"}
or as specified using \code{sf::st_crs()} like \code{sf::st_crs(4326)}
the default value.}

\item{na_rm}{a logical value. If \code{TRUE} remove sites with incomplete
coordinates. Default is \code{FALSE}.}
}
\value{
An \code{sf} object with a \code{site} column specifying site coordinates.
}
\description{
Convert a flat \code{data.frame} with site coordinates into a proper \code{sf} object
that can then be used by other functions. This function assumes that
the coordinates are given in WGS84 (longitude vs. latitude). The function
automatically removes repeated coordinates from the input dataset.
}
\examples{
filename <- system.file(
    "extdata", "woodiv_raw_data.csv",
    package = "funbiogeo"
)
all_data <- read.csv(filename)

head(all_data)

site_locations <- fb_format_site_locations(
    all_data, "site", "longitude",
    "latitude"
)
head(site_locations)
}
