% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_format_site_species.R
\name{fb_format_site_species}
\alias{fb_format_site_species}
\title{Format site x species Object from Long Data}
\usage{
fb_format_site_species(data, site, species, value, na_to_zero = TRUE)
}
\arguments{
\item{data}{a \code{data.frame} in a long format (see example).}

\item{site}{a character of length 1. Name of the column with site labels.}

\item{species}{a character of length 1. Name of the column with species
names.}

\item{value}{a character of length 1. Name of the column with species
occurrence/abundance.}

\item{na_to_zero}{a logical value. If \code{TRUE} (default) NA are replaced by
\code{0}.}
}
\value{
A data.frame with sites in rows and species in columns. The first
column is named \code{"site"} and contains the name of the sites.
}
\description{
...
}
\examples{
library("funbiogeo")

filename <- system.file("extdata", "raw_mammals_data.csv", 
                        package = "funbiogeo")
all_data <- read.csv(filename)

head(all_data)

site_species <- fb_format_site_species(all_data, "site", "species", "count")
head(site_species)
}
