% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_format_species_categories.R
\name{fb_format_species_categories}
\alias{fb_format_species_categories}
\title{Extract species x categories information from long format data}
\usage{
fb_format_species_categories(data, species, category)
}
\arguments{
\item{data}{a \code{data.frame} in a long format (see example).}

\item{species}{a \code{character} of length 1. Name of the column with species
names.}

\item{category}{a \code{character} of length 1. Name of the column with species
category.}
}
\value{
A \code{data.frame} with species in rows and two columns (species name and
species category).
}
\description{
Convert a flat \code{data.frame} with species names and species (supra-)category
(e.g. family, order, endemism status, etc.) into a proper \code{data.frame} object
that can then be used by other functions.
The final output contains species in rows and two columns (species name and
species category).
}
\examples{
library("funbiogeo")

filename <- system.file("extdata", "funbiogeo_raw_data.csv", 
                        package = "funbiogeo")
all_data <- read.csv(filename)

head(all_data)

species_categories <- fb_format_species_categories(all_data, "species", 
                                                   "genus")
species_categories[1:6, ]
}
