% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_format_species_traits.R
\name{fb_format_species_traits}
\alias{fb_format_species_traits}
\title{Format species x traits Object from Long Data}
\usage{
fb_format_species_traits(data_long, species, traits)
}
\arguments{
\item{data_long}{a \code{data.frame} in a long format (see example).}

\item{species}{a character of length 1. Name of the column with species
names.}

\item{traits}{a character of length >= 1. Name(s) of trait column(s).}
}
\value{
A \code{data.frame} with species in rows and traits in columns, with the
first column names \code{"species"} containing the species names
}
\description{
...
}
\examples{
library("funbiogeo")

filename <- system.file("extdata", "raw_mammals_data.csv", 
                        package = "funbiogeo")
all_data <- read.csv(filename)

head(all_data)

traits <- c("adult_body_mass", "gestation_length", "litter_size", 
            "max_longevity", "sexual_maturity_age", "diet_breadth")

species_traits <- fb_format_species_traits(all_data, "species", traits)
species_traits
}
