% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_format_species_traits.R
\name{fb_format_species_traits}
\alias{fb_format_species_traits}
\title{Format species x traits object}
\usage{
fb_format_species_traits(data, species, traits)
}
\arguments{
\item{data}{a \code{data.frame} in a long format (see example).}

\item{species}{a character of length 1. Name of the column with species
names.}

\item{traits}{a character of length >= 1. Name(s) of trait column(s).}
}
\value{
A \code{data.frame} with species in rows and traits in columns.
}
\description{
...
}
\examples{
library("funbiogeo")

filename <- system.file("extdata", "raw_trees_data.csv", 
                        package = "funbiogeo")
all_data <- read.csv2(filename)

head(all_data)

species_traits <- fb_format_species_traits(all_data, "species", 
                                           c("sla", "n_mass", "wood_dens"))
species_traits
}
