% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_get_all_trait_coverages_by_site.R
\name{fb_get_all_trait_coverages_by_site}
\alias{fb_get_all_trait_coverages_by_site}
\title{Compute Trait Coverage per Site for Each Trait}
\usage{
fb_get_all_trait_coverages_by_site(
  site_species,
  species_traits,
  all_traits = TRUE
)
}
\arguments{
\item{site_species}{a \code{data.frame} with sites in rows and species in columns.
\strong{NOTE}: the first column should be named \strong{\code{"site"}} and indicate site
names. The other columns should be named according to species names.}

\item{species_traits}{a \code{data.frame} with species in rows and
traits as columns. \strong{NOTE}: The first column should be named \strong{\code{"species"}}
and contain species names. The other columns should be named according
to trait names.}

\item{all_traits}{a logical (default = \code{TRUE}) which tell if the coverage
considering all provided traits should be provided in an additional column
\code{all_traits}}
}
\value{
a data.frame with a column with sites and one column per provided
trait giving its coverage (percent species per site, weighted by abundance
that have trait data), and, when argument \code{all_traits = TRUE},
an additional column named \code{all_traits} considering the coverage of all
traits taken together.
}
\description{
Compute trait coverage for all sites, i.e., the percentage of total
abundance/presence of species that have traits data compared to total
species. This function assumes that all species provided in the traits
dataset have all their traits specified (meaning that all species have either
known or \code{NA} values reported as their traits).
The coverage of each trait separately is returned as well as all traits taken
together if wanted.
}
\examples{
site_trait_cov <- fb_get_all_trait_coverages_by_site(
  woodiv_site_species, woodiv_traits
)

head(site_trait_cov)
}
