% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_get_all_trait_coverages_by_site.R
\name{fb_get_all_trait_coverages_by_site}
\alias{fb_get_all_trait_coverages_by_site}
\title{Compute Trait Coverage per Site for All Traits}
\usage{
fb_get_all_trait_coverages_by_site(site_species, species_traits)
}
\arguments{
\item{site_species}{a \code{data.frame} with sites in rows and species in columns.
\strong{NOTE}: the first column should be named \strong{\code{"site"}} and indicate site
names. The other columns should be named according to species names.}

\item{species_traits}{a \code{data.frame} with species in rows and
traits as columns. \strong{NOTE}: The first column should be named \strong{\code{"species"}}
and contain species names. The other columns should be named according
to trait names.}
}
\value{
a data.frame with a column with sites and one column per provided
trait giving its coverage (percent species per site, weighted by abundance
that have trait data)
}
\description{
Compute Trait Coverage per Site for All Traits
}
\examples{
fb_get_all_trait_coverages_by_site(site_species, species_traits)
}
