% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_get_environment.R
\name{fb_get_environment}
\alias{fb_get_environment}
\title{Extract average raster values at sites locations}
\usage{
fb_get_environment(sites_locations, environment_raster)
}
\arguments{
\item{sites_locations}{an \code{sf} object with each sites defined as points}

\item{environment_raster}{a \code{raster} object of environmental values}
}
\value{
A \code{data.frame} with average environmental values (columns) per site
(rows).
}
\description{
Extract average raster values at sites locations
}
\examples{
library("funbiogeo")

data("sites_locs")

## Convert data frame to sf object ----
sites_locs_sf <- sf::st_as_sf(sites_locs, 
                              coords = c("longitude", "latitude"))

## Import climate rasters ----
prec <- system.file("extdata", "annual_tot_prec.tif", package = "funbiogeo")
tavg <- system.file("extdata", "annual_mean_temp.tif", package = "funbiogeo")

layers <- terra::rast(c(tavg, prec))

## Extract environment at sites ----
sites_env <- fb_get_environment(sites_locs_sf, layers)

## Add sites labels ----
sites_env$"ID" <- rownames(sites_locs)
}
