% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_get_environment.R
\name{fb_get_environment}
\alias{fb_get_environment}
\title{Extract average raster values at sites locations}
\usage{
fb_get_environment(site_locations, environment_raster)
}
\arguments{
\item{site_locations}{an \code{sf} object with each sites defined as points}

\item{environment_raster}{a \code{raster} object of environmental values}
}
\value{
a \code{data.frame} with average environmental values per site
}
\description{
Extract average raster values at sites locations
}
