% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_get_environment.R
\name{fb_get_environment}
\alias{fb_get_environment}
\title{Extract average raster values at sites locations}
\usage{
fb_get_environment(sites_locations, environment_raster)
}
\arguments{
\item{sites_locations}{an \code{sf} object with each sites defined as points}

\item{environment_raster}{a \code{raster} object of environmental values}
}
\value{
A \code{data.frame} with average environmental values (columns) per site
(rows).
}
\description{
Extract average raster values at sites locations
}
\examples{
## Add an example ----
}
