% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_get_environment.R
\name{fb_get_environment}
\alias{fb_get_environment}
\title{Extract average raster values at sites locations}
\usage{
fb_get_environment(site_locations, environment_raster)
}
\arguments{
\item{site_locations}{an \code{sf} object with all sites.}

\item{environment_raster}{a \code{SpatRaster} object (package \code{terra}).
A single or multi-layers environmental raster.}
}
\value{
A \code{data.frame} with average environmental values (columns) per site
(rows).
}
\description{
Extract average raster values at sites locations
}
\examples{
library("funbiogeo")

data("site_locs")

## Import climate rasters ----
prec <- system.file("extdata", "annual_tot_prec.tif", package = "funbiogeo")
tavg <- system.file("extdata", "annual_mean_temp.tif", package = "funbiogeo")

layers <- terra::rast(c(tavg, prec))

}
