% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_get_trait_combination_coverage.R
\name{fb_get_trait_combination_coverage}
\alias{fb_get_trait_combination_coverage}
\title{Compute site trait coverage for each trait combination}
\usage{
fb_get_trait_combination_coverage(
  site_species,
  species_traits,
  comb_size = NULL
)
}
\arguments{
\item{site_species}{a \code{data.frame} with sites in rows and species in columns.
\strong{NOTE}: the first column should be named \strong{\code{"site"}} and indicate site
names. The other columns should be named according to species names.}

\item{species_traits}{a \code{data.frame} with species in rows and
traits as columns. \strong{NOTE}: The first column should be named \strong{\code{"species"}}
and contain species names. The other columns should be named according
to trait names.}

\item{comb_size}{an integer vector defining one or more sizes of combinations
(default: \code{NULL})}
}
\value{
a data.frame with the following columns:
\itemize{
\item \code{site} with the site indices from \code{site_species},
\item \code{combination_length} with the number of traits in given combinations,
\item \code{combination_name} with the name of the trait combination (concatenated
trait names with \verb{__}),
\item \code{trait_coverage} the corresponding trait coverage for the given
trait combination and site.
}
}
\description{
This function computes trait coverage for each site for different trait
combinations. If not provided, consider all possible trait combinations.
The function will not run if the total number of combinations given is over
10,000.
}
\examples{
# Compute Coverages using All Trait Combinations
all_combinations = fb_get_trait_combination_coverage(
site_species, species_traits
)

# Get only combinations of 3 traits
three_traits = fb_get_trait_combination_coverage(
site_species, species_traits, 3
)

# Combinations of 2, 3, and 4 traits
two_to_four = fb_get_trait_combination_coverage(
site_species, species_traits, c(2, 3, 4)
)
}
