% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_get_trait_coverage_by_site.R
\name{fb_get_trait_coverage_by_site}
\alias{fb_get_trait_coverage_by_site}
\title{Compute traits coverage}
\usage{
fb_get_trait_coverage_by_site(site_species, species_traits)
}
\arguments{
\item{site_species}{a \code{data.frame} with sites in rows and species in columns.
The first column should be named \code{"site"} and contain site names. The other
columns should be named according to species names.}

\item{species_traits}{a \code{data.frame} with species in rows and
traits as columns. The first column should be named \code{"species"}
and contain species names. The other columns should be named according
to trait names.}
}
\value{
A \code{data.frame} with \code{n} rows (where \code{n} is the number of sites) and
two columns: \code{site}, the site label, and \code{trait_coverage}, the percent of
total abundance/presence of species that have traits data.
}
\description{
Compute traits coverage for all sites, i.e., the percentage of total
abundance/presence of species that have traits data compared to total
species.
This function assumes that all species provided in the traits dataset have
all their traits specified.
}
\examples{
library("funbiogeo")

data("site_species")
data("species_traits")

cover <- fb_get_trait_coverage_by_site(site_species, species_traits)
}
