% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_make_report.R
\name{fb_make_report}
\alias{fb_make_report}
\title{Create an R Markdown report}
\usage{
fb_make_report(
  path = ".",
  filename = NULL,
  title = NULL,
  author = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{path}{a \code{character} of length 1. The directory in which the \code{.Rmd}
file will be created. This directory must exist.}

\item{filename}{a \code{character} of length 1. The name of the \code{.Rmd} file to be
created. If \code{NULL} (default) the \code{.Rmd} file will be named
\code{funbiogeo_report.Rmd}.}

\item{title}{a \code{character} of length 1. The title of the report.
If \code{NULL} (default) the title will be \verb{funbiogeo Report}.}

\item{author}{a \code{character} of length 1. The author(s) of the report.
If \code{NULL} (default) no author will be added.}

\item{overwrite}{a logical. If this file is already present and
\code{overwrite = TRUE}, it will be erased and replaced by the template.
Default is \code{FALSE}.}
}
\value{
No return value.
}
\description{
Creates an R Markdown (\code{.Rmd}) report from a template to explore and
summarize users data. User can modify this report and use the function
\code{\link[rmarkdown:render]{rmarkdown::render()}} to convert this \code{.Rmd} in different formats:
\itemize{
\item HTML document (\code{output_format = "html_document"});
\item PDF document (\code{output_format = "pdf_document"});
\item WORD document (\code{output_format = "word_document"});
\item HTML, PDF and WORD documents (\code{output_format = "all"}).
}
}
\examples{
# Create temporary folder ----
temp_path <- tempdir()

# Create report ----
fb_make_report(path = temp_path, author = "Casajus N. and Grenie M.")

\dontrun{
# Open Rmd file ----
utils::file.edit(file.path(temp_path, "funbiogeo_report.Rmd"))
}
}
