% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_make_report.R
\name{fb_make_report}
\alias{fb_make_report}
\title{Create an R Markdown Report on Site/Species/Trait Coverage with Plots}
\usage{
fb_make_report(
  path = ".",
  filename = NULL,
  title = NULL,
  author = NULL,
  species_traits_name,
  site_species_name,
  site_locations_name,
  overwrite = FALSE,
  open = TRUE
)
}
\arguments{
\item{path}{a \code{character} of length 1. The directory in which the \code{.Rmd}
file will be created. This directory must exist.}

\item{filename}{a \code{character} of length 1. The name of the \code{.Rmd} file to be
created. If \code{NULL} (default) the \code{.Rmd} file will be named
\code{funbiogeo_report.Rmd}.}

\item{title}{a \code{character} of length 1. The title of the report.
If \code{NULL} (default) the title will be \verb{funbiogeo Report}.}

\item{author}{a \code{character} of length 1. The author(s) of the report.
If \code{NULL} (default) no author will be added.}

\item{species_traits_name}{a \code{character} of length 1. The \strong{name} of the
species x traits dataset (not the object). Note that before rendering the
report this dataset must be loaded.}

\item{site_species_name}{a \code{character} of length 1. The \strong{name} of the
sites x species dataset (not the object). Note that before rendering the
report this dataset must be loaded.}

\item{site_locations_name}{a \code{character} of length 1. The \strong{name} of the
sites x locations dataset (not the object). Note that before rendering the
report this dataset must be loaded.}

\item{overwrite}{a logical. If this file is already present and
\code{overwrite = TRUE}, it will be erased and replaced by the template.
Default is \code{FALSE}.}

\item{open}{a logical. If \code{TRUE} (default), this file will be opened on the
text editor.}
}
\value{
No return value.
}
\description{
Creates an R Markdown (\code{.Rmd}) report from a template to explore and
summarize users data. User can modify this report and use the function
\code{\link[rmarkdown:render]{rmarkdown::render()}} to convert this \code{.Rmd} in different formats:
\itemize{
\item HTML document (\code{output_format = "bookdown::html_document2"});
\item PDF document (\code{output_format = "bookdown::pdf_document2"});
\item Word document (\code{output_format = "bookdown::word_document2"});
\item HTML, PDF and Word documents (\code{output_format = "all"}).
}
}
\examples{
# Create temporary folder
temp_path <- tempdir()

# Load data
data("species_traits")
data("site_species")
data("site_locations")

# Create report
fb_make_report(
  path                = temp_path, 
  author              = "Casajus N. and Grenié M.",
  species_traits_name = "species_traits",
  site_species_name   = "site_species",
  site_locations_name = "site_locations",
  open = FALSE
)

\dontrun{
# Open Rmd file ----
utils::file.edit(file.path(temp_path, "funbiogeo_report.Rmd"))

# Render Rmd file ----
rmarkdown::render(file.path(temp_path, "funbiogeo_report.Rmd"), 
                  output_format = "all")
}
}
