% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_map_site_data.R
\name{fb_map_site_data}
\alias{fb_map_site_data}
\title{Map Arbitrary Site Data}
\usage{
fb_map_site_data(site_locations, site_data, selected_col, background = FALSE)
}
\arguments{
\item{site_locations}{an \code{sf} object with the spatial geometries of sites.
\strong{NOTE}: the first column should be named \strong{\code{"site"}}
and indicate site names.}

\item{site_data}{\code{data.frame()} of additional site information containing
the column \code{"site"} to merge with the \code{site_locations} argument}

\item{selected_col}{\code{character(1)} name of the column to plot}

\item{background}{a \code{logical}. If \code{TRUE} adds a layer of country boundaries
from Natural Earth.}
}
\value{
a \code{ggplot} object.
}
\description{
From the site-locations data and a dataset organized by site, plot a map of
this information.
The returned plot is as little customized as possible to let the user do
the customization.
}
\examples{
site_rich <- fb_count_species_by_site(woodiv_site_species)

# Map of Species Richness
rich_map <- fb_map_site_data(woodiv_locations, site_rich, "n_species")
rich_map

# Customize the map
rich_map +
  ggplot2::scale_fill_viridis_c("Species Richness")

# Map w/ a background
fb_map_site_data(
  woodiv_locations, site_rich, "n_species", background = TRUE
)
}
