% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_map_site_traits_completeness.R
\name{fb_map_site_traits_completeness}
\alias{fb_map_site_traits_completeness}
\title{Map Trait Coverage Per Site}
\usage{
fb_map_site_traits_completeness(
  site_locations,
  site_species,
  species_traits,
  all_traits = TRUE,
  background = FALSE
)
}
\arguments{
\item{site_locations}{an \code{sf} object with the spatial geometries of sites.
\strong{NOTE}: the first column should be named \strong{\code{"site"}}
and indicate site names.}

\item{site_species}{a \code{data.frame} with sites in rows and species in columns.
\strong{NOTE}: the first column should be named \strong{\code{"site"}} and indicate site
names. The other columns should be named according to species names.}

\item{species_traits}{a \code{data.frame} with species in rows and
traits as columns. \strong{NOTE}: The first column should be named \strong{\code{"species"}}
and contain species names. The other columns should be named according
to trait names.}

\item{all_traits}{a logical (default = \code{TRUE}) which tell if the coverage
considering all provided traits should be provided in an additional column
\code{all_traits}}

\item{background}{a \code{logical}. If \code{TRUE} adds a layer of country boundaries
from Natural Earth.}
}
\value{
a 'ggplot2' object
}
\description{
Returns a \code{ggplot2} map of sites colored by trait coverage (proportion
of species having a known trait value). By default shows one plot for each
trait and add an additional facet named \code{"all_traits"} considering the
trait coverage with all traits taken together.
}
\examples{
fb_map_site_traits_completeness(
    woodiv_locations, woodiv_site_species, woodiv_traits
)
}
