% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_plot_distribution_site_trait_coverage.R
\name{fb_plot_distribution_site_trait_coverage}
\alias{fb_plot_distribution_site_trait_coverage}
\title{Plot Distribution of Trait Coverages across all Sites}
\usage{
fb_plot_distribution_site_trait_coverage(
  site_species,
  species_traits,
  all_traits = TRUE
)
}
\arguments{
\item{site_species}{a \code{data.frame} with sites in rows and species in columns.
\strong{NOTE}: the first column should be named \strong{\code{"site"}} and indicate site
names. The other columns should be named according to species names.}

\item{species_traits}{a \code{data.frame} with species in rows and
traits as columns. \strong{NOTE}: The first column should be named \strong{\code{"species"}}
and contain species names. The other columns should be named according
to trait names.}

\item{all_traits}{a logical (default = \code{TRUE}) which tell if the coverage
considering all provided traits should be provided in an additional column
\code{all_traits}}
}
\value{
a 'ggplot2' object
}
\description{
Plots the distributions of trait coverage across site, i.e. the proportion of
species weighted by abundance with known trait values, for each trait
separately and all traits taken together.
The trait distributions are ordered from the lowest to the highest average
trait coverage (top to bottom). The top always displays a distribution named
\code{"all_traits"} that contains the distribution of coverage all traits taken
together.
}
\examples{
fb_plot_distribution_site_trait_coverage(site_species, species_traits)

}
