% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_plot_distribution_site_trait_coverage.R
\name{fb_plot_distribution_site_trait_coverage}
\alias{fb_plot_distribution_site_trait_coverage}
\title{Distribution of Trait Coverages across all sites}
\usage{
fb_plot_distribution_site_trait_coverage(site_species, species_traits)
}
\arguments{
\item{site_species}{a \code{data.frame} with sites in rows and species in columns.
\strong{NOTE}: the first column should be named \strong{\code{"site"}} and indicate site
names. The other columns should be named according to species names.}

\item{species_traits}{a \code{data.frame} with species in rows and
traits as columns. \strong{NOTE}: The first column should be named \strong{\code{"species"}}
and contain species names. The other columns should be named according
to trait names.}
}
\value{
a 'ggplot2' object
}
\description{
Distribution of Trait Coverages across all sites
}
\examples{
fb_plot_distribution_site_trait_coverage(site_species, species_traits)

}
