% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_plot_number_species_by_trait.R
\name{fb_plot_number_species_by_trait}
\alias{fb_plot_number_species_by_trait}
\title{Plot Number of Species per Trait}
\usage{
fb_plot_number_species_by_trait(
  species_traits,
  species_categories = NULL,
  threshold_species_proportion = NULL
)
}
\arguments{
\item{species_traits}{a \code{data.frame} with species in rows and
traits as columns. \strong{NOTE}: The first column should be named \strong{\code{"species"}}
and contain species names. The other columns should be named according
to trait names.}

\item{species_categories}{(default = \code{NULL}) 2-columns \code{data.frame} giving
species categories, with the first column describing the species name, and
the second column giving their corresponding categories}

\item{threshold_species_proportion}{\code{numeric(1)} [default = \code{NULL}]\cr{}
between 0 and 1. The percentage of species coverage threshold.}
}
\value{
a ggplot2 object
}
\description{
Display a lollipop graph showing the number and proportion of species
with non-NA trait for each trait ranked in decreasing order.
}
\examples{
data(species_traits)

fb_plot_number_species_by_trait(species_traits)

# Add a vertical cutoff line (12.5\% of species)
fb_plot_number_species_by_trait(species_traits, NULL, 1/8)

}
