% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_plot_site_traits_completeness.R
\name{fb_plot_site_traits_completeness}
\alias{fb_plot_site_traits_completeness}
\title{Plot Sites per Traits Completeness}
\usage{
fb_plot_site_traits_completeness(site_species, species_traits)
}
\arguments{
\item{site_species}{a \code{data.frame} with sites in rows and species in columns.
The first column should be named \code{"site"} and contain site names. The other
columns should be named according to species names.}

\item{species_traits}{a \code{data.frame} with species in rows and
traits as columns. The first column should be named \code{"species"}
and contain species names. The other columns should be named according
to trait names.}
}
\value{
a ggplot2 object
}
\description{
Plot Sites per Traits Completeness
}
\examples{
\dontrun{fb_plot_site_traits_completeness(site_species, species_traits)}
}
