% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_plot_site_traits_completeness.R
\name{fb_plot_site_traits_completeness}
\alias{fb_plot_site_traits_completeness}
\title{Plot Trait Coverage per Site for each Trait}
\usage{
fb_plot_site_traits_completeness(
  site_species,
  species_traits,
  species_categories = NULL,
  all_traits = TRUE
)
}
\arguments{
\item{site_species}{a \code{data.frame} with sites in rows and species in columns.
\strong{NOTE}: the first column should be named \strong{\code{"site"}} and indicate site
names. The other columns should be named according to species names.}

\item{species_traits}{a \code{data.frame} with species in rows and
traits as columns. \strong{NOTE}: The first column should be named \strong{\code{"species"}}
and contain species names. The other columns should be named according
to trait names.}

\item{species_categories}{(default = \code{NULL}) 2-columns \code{data.frame} giving
species categories, with the first column describing the species name, and
the second column giving their corresponding categories}

\item{all_traits}{a logical (default = \code{TRUE}) which tell if the coverage
considering all provided traits should be provided in an additional column
\code{all_traits}}
}
\value{
a ggplot2 object
}
\description{
Display a binary heatmap visualizing the site x traits matrix with colors
displaying the proportion of occurring species with known trait values.
Traits are ordered from the most to the least known (left to right).
Sites are ordered from the ones with highest to lowest overall trait coverage
(bottom to top). The site average proportion of species with known trait
for each trait (across all sites) is shown in the x-axis labels.
An additional column at the very right of the plot named \code{"all_traits"}
shows a summary considering traits together.
}
\examples{
fb_plot_site_traits_completeness(woodiv_site_species, woodiv_traits)

}
