% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_plot_species_traits_completeness.R
\name{fb_plot_species_traits_completeness}
\alias{fb_plot_species_traits_completeness}
\title{Plot Trait Coverage per Species for each Trait}
\usage{
fb_plot_species_traits_completeness(
  species_traits,
  species_categories = NULL,
  all_traits = TRUE
)
}
\arguments{
\item{species_traits}{a \code{data.frame} with species in rows and
traits as columns. \strong{NOTE}: The first column should be named \strong{\code{"species"}}
and contain species names. The other columns should be named according
to trait names.}

\item{species_categories}{2-columns \code{data.frame} giving species categories
\code{NULL} by default, with the first column describing the species name, and
the second column giving their corresponding categories}

\item{all_traits}{a logical (default = \code{TRUE}) which tell if the coverage
considering all provided traits should be provided in an additional column
\code{all_traits}}
}
\value{
a \code{ggplot2} object
}
\description{
Display a binary heatmap visualizing the species x traits matrix with colors
displaying present and missing traits. Traits are ordered from the most to
the least known (left to right).
Species are ordered from the ones with most to the ones with least traits
(bottom to top). The proportion of species with non-missing traits is shown
on the x-axis labels.
An additional column at the very right of the plot named \code{"all_traits"}
shows a summary considering if all other traits are known.
}
\examples{
fb_plot_species_traits_completeness(woodiv_traits)

}
