% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_plot_species_traits_completeness.R
\name{fb_plot_species_traits_completeness}
\alias{fb_plot_species_traits_completeness}
\title{Show Species by Trait Completeness}
\usage{
fb_plot_species_traits_completeness(species_traits)
}
\arguments{
\item{species_traits}{a \code{data.frame} with species in rows and
traits as columns. The first column should be named \code{"species"}
and contain species names. The other columns should be named according
to trait names.}
}
\value{
a \code{ggplot2} object
}
\description{
Show Species by Trait Completeness
}
\examples{
data(species_traits)
\dontrun{fb_plot_species_traits_completeness(species_traits)}

}
