% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funbiogeo-package.R
\docType{package}
\name{funbiogeo-package}
\alias{funbiogeo}
\alias{funbiogeo-package}
\title{funbiogeo: Streamlining Functional Biogeography Analyses}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Helps users with analyses in functional biogeography by loading and combining data, computing trait coverage, as well as computing functional diversity indices, drawing maps, correlating them with the environment, and upscaling assemblages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://frbcesab.github.io/funbiogeo/}
  \item \url{https://github.com/frbcesab/funbiogeo}
  \item Report bugs at \url{https://github.com/frbcesab/funbiogeo/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]
  \item Matthias Grenié \email{matthias.grenie@univ-grenoble-alpes.fr} (\href{https://orcid.org/0000-0002-4659-7522}{ORCID})
}

}
\keyword{internal}
