% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funbiogeo-package.R
\docType{package}
\name{funbiogeo-package}
\alias{funbiogeo}
\alias{funbiogeo-package}
\title{funbiogeo: Functional Biogeography}
\description{
Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do 
    eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim 
    veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea  
    commodo consequat.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/frbcesab/funbiogeo}
  \item Report bugs at \url{https://github.com/frbcesab/funbiogeo/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID})

Authors:
\itemize{
  \item Matthias Grénié \email{matthias.grenie@idiv.de}
  \item Kévin Cazelles \email{kevin.cazelles@gmail.com}
}

}
\keyword{internal}
