% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{site_locations}
\alias{site_locations}
\title{Sites x Locations of Western Europe mammals occurrences}
\format{
An \code{sf} object with 1,505 rows (grid cells) with a WGS84 (EPSG:4236)
Coordinate Reference System.
}
\usage{
site_locations
}
\description{
This dataset contains spatial coordinates (grid cells of 0.5° x 0.5°
horizontal resolution) of sites of Western Europe where mammals species
presence/absence (see \code{site_species} dataset) has been sampled.

See \url{https://github.com/frbcesab/eumammals} for a complete description
of the workflow used to create this dataset.

This dataset shows the format of the argument \code{site_locations} used in
several functions of \code{funbiogeo}.

The variable \code{site} corresponds to the sites labels.
}
\examples{
data("site_locations")
class(site_locations)
dim(site_locations)
head(site_locations)
}
\keyword{datasets}
