% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{site_species}
\alias{site_species}
\title{Data: sites x species object}
\format{
A \code{data.frame} with 1,505 rows (sites) and 149 columns (species
occurrence).
}
\usage{
site_species
}
\description{
This dataset contains the presence/absence of mammals species in Western
Europe. This dataset is derived from IUCN range maps downloaded at:
https://www.iucnredlist.org/resources/spatial-data-download. In respect to
the license, species names have been anonymized.

See https://github.com/frbcesab/eumammals for a complete description of the
workflow used to create this dataset.

This dataset shows the format of the argument \code{site_species} used in
several functions of \code{funbiogeo}.

Note that sites labels are specified in row names.
}
\examples{
data("site_species")
class(site_species)
dim(site_species)
site_species[1:6, 1:6]
}
\references{
IUCN (2021) The IUCN Red List of Threatened Species. Version 2021-2.
Downloaded on 2021/10/11.
}
\keyword{datasets}
