% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{woodiv_locations}
\alias{woodiv_locations}
\title{Sites x Locations of some Mediterranean Conifers}
\format{
An \code{sf} \code{POLYGON} object with 5,366 rows (grid cells) defined in the
ETRS89-extended / LAEA Europe Coordinate Reference System (EPSG:3035).
}
\usage{
woodiv_locations
}
\description{
This dataset is derived from the WOODIV database (available at:
\url{https://www.nature.com/articles/s41597-021-00873-3}). It contains the
grid cells of sites (10 km x 10 km horizontal
resolution) sampled in Portugal, Spain, France, and Italy (Mediterranean
part) for which at least one of the 24 conifer tree species occurs.

This dataset exemplifies the argument \code{site_locations} used in
several functions of \code{funbiogeo}. The variable \code{site} corresponds to the
site labels (must match the same column in \code{site_species}) and the variable
\code{country} will be used as a site grouping factor.
}
\examples{
data("woodiv_locations")
class(woodiv_locations)
dim(woodiv_locations)
head(woodiv_locations)
}
\references{
Monnet AC, Cilleros K, Médail F \emph{et al.} (2021) WOODIV, a database of
occurrences, functional traits, and phylogenetic data for all
Euro-Mediterranean trees. \strong{Scientific Data}, 8, 89.
DOI: \doi{10.1038/s41597-021-00873-3}.
}
\keyword{datasets}
