% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{woodiv_site_species}
\alias{woodiv_site_species}
\title{Sites x Species of some Mediterranean Conifers}
\format{
A \code{data.frame} with 5,366 rows (sites) and 25 columns (1 column for
site label and 24 for species occurrence).
}
\usage{
woodiv_site_species
}
\description{
This dataset is derived from the WOODIV database (available at:
\url{https://www.nature.com/articles/s41597-021-00873-3}). It contains the
presence/absence of the 24 Conifer tree species occurring in Portugal,
Spain, France, and Italy (Mediterranean part). The presence/absence is
available at 5,366 sites (grid cells of 10 x 10 km horizontal resolution).
Sites coordinates are provided in \code{site_locations}.

This dataset shows the format of the argument \code{site_species} used in
several functions of \code{funbiogeo}. Note that site labels must be the first
column and must be named \code{site}.
}
\examples{
data("woodiv_site_species")
class(woodiv_site_species)
dim(woodiv_site_species)
woodiv_site_species[1:6, 1:6]
}
\references{
Monnet AC, Cilleros K, Médail F \emph{et al.} (2021) WOODIV, a database of
occurrences, functional traits, and phylogenetic data for all
Euro-Mediterranean trees. \strong{Scientific Data}, 8, 89.
DOI: \href{https://doi.org/10.1038/s41597-021-00873-3}{10.1038/s41597-021-00873-3}
}
\keyword{datasets}
