% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{woodiv_traits}
\alias{woodiv_traits}
\title{Species x Traits of some Mediterranean Conifers}
\format{
A \code{data.frame} with 24 rows (species) and the following five columns:
\describe{
\item{species}{species name corresponding to the columns of \code{site_species}}
\item{plant_height}{adult plant height (in \eqn{m})}
\item{seed_mass}{seed mass (in \eqn{g})}
\item{sla}{specific leaf area, i.e. the ratio between leaf area and dry
mass (in \eqn{m^{2}\cdot kg^{-1}})}
\item{wood_density}{wood density, i.e. stem specific density
(in \eqn{kg\cdot dm^{–3}})}
}
}
\usage{
woodiv_traits
}
\description{
This dataset is derived from the WOODIV database (available at:
\url{https://www.nature.com/articles/s41597-021-00873-3}). It contains the
values of four functional traits for 24 Conifer tree species occurring in
Portugal, Spain, France, and Italy (Mediterranean part).

This dataset shows the format of the argument \code{species_traits} used in
several functions of \code{funbiogeo}. Note that species names must be
the first column and must be named \code{species}.
}
\examples{
data("woodiv_traits")
class(woodiv_traits)
dim(woodiv_traits)
head(woodiv_traits)
}
\references{
Monnet AC, Cilleros K, Médail F \emph{et al.} (2021) WOODIV, a database of
occurrences, functional traits, and phylogenetic data for all
Euro-Mediterranean trees. \strong{Scientific Data}, 8, 89.
DOI: \href{https://doi.org/10.1038/s41597-021-00873-3}{10.1038/s41597-021-00873-3}
}
\keyword{datasets}
